added a expand all option to the recipe
