+++ docs/data/data-grid/row-recipes/DetailPanelExpandCollapseAll.tsx.preview
@@ -1,0 +1,6 @@
+<DataGridPro
+  rows={rows}
+  columns={columns}
+  getDetailPanelContent={getDetailPanelContent}
+  getDetailPanelHeight={getDetailPanelHeight}
+/>
