+++ docs/data/data-grid/row-recipes/row-recipes.md
@@ -14,14 +14,15 @@ However, you can [control the expanded detail panels](/x/react-data-grid/master-
 
 {{"demo": "DetailPanelOneExpandedRow.js", "bg": "inline", "defaultCodeOpen": false}}
 
-## Collapse all detail panels with a custom header
+## Expand or collapse all detail panels with a custom header
 
-This feature allows you to create a custom header element that collapses all currently expanded detail panels at once.
+This feature allows you to create a custom header element that expands or collapses all detail panels at once.
 
 Here is how it works:
 
-- Clicking the custom header triggers a collapse action on all open detail panels.
+- Clicking the custom header triggers a expand/collapse action on all detail panels.
 - It checks the status of open panels using the [`useGridSelector` hook](/x/react-data-grid/state/#with-usegridselector) to access the grid's state.
-- The example leverages [`setExpandedDetailPanels`](/x/api/data-grid/grid-api/#grid-api-prop-setExpandedDetailPanels) from the [Grid API](/x/react-data-grid/api-object/#how-to-use-the-api-object).
+- For getting all rows that have a detail panel it uses the [`useGridSelector` hook](/x/react-data-grid/state/#with-usegridselector) as well.
+- The example then leverages [`setExpandedDetailPanels`](/x/api/data-grid/grid-api/#grid-api-prop-setExpandedDetailPanels) from the [Grid API](/x/react-data-grid/api-object/#how-to-use-the-api-object) to expand or collapse all detail panels.
 
-{{"demo": "DetailPanelCollapseAll.js", "bg": "inline", "defaultCodeOpen": false}}
+{{"demo": "DetailPanelExpandCollapseAll.js", "bg": "inline", "defaultCodeOpen": false}}
